Remove Restriction on Empty Table Payloads (#350)
This is a quick follow-up to the previous PR. There are scenarios
(scripts and certain use cases), where it’s valid for empty tables to be
included as payload items. To accommodate this, I’ve removed the earlier
validation check that rejected empty tables.

Additionally, the corresponding test case that enforced this validation
has been removed.