Move print statements to logger warnings (#423)
Put warning messages in the logger instead or print statements. This
should fix issues from the automated testing machines, that fail due to
these messages even though the tests pass.

Most of these are always triggered at start up, but they are innocuous.
The ones that are actual problems to look into are still under the
print_allowed() condition, so they pop up to the user that is using the
UI tools.