Add XSS Injection Check for Trees and Tables Using Bleach (#421)
- **Tree Payloads:** Added validation to sanitize tree node values using
bleach. If the sanitized value differs from the original, an error is
raised.
- **Table Payloads:** Added validation to sanitize string elements in
table arrays using bleach.clean(). If the sanitized value differs from
the original, an error is raised.
- **Environment Variable:** Introduced the ADR_VALIDATION_BETAFLAG_ANSYS
environment variable to enable or disable this feature. This ensures the
changes are only active for testing purposes and do not impact end
users.

PR Story Item:
https://tfs.ansys.com:8443/tfs/ANSYS_Development/Portfolio/_boards/board/t/Nexus/Stories/?workitem=1345806